Streamlit_NET_app/requirements.txt
scikit-learn==1.0.2
streamlit==1.8.0
opencv-python-headless==4.5.5.62
matplotlib==3.5.1
tensorflow==2.7.0
Pillow==8.4.0
plotly==5.5.0
numpy==1.21.4
pandas==1.3.4
